+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableByteArray.kt
@@ -671,6 +671,13 @@ public value class ImmutableByteArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns the first element having the largest value according to the provided [comparator] or null if empty.
+     */
+    public fun maxWithOrNull(comparator: Comparator<in Byte>): Byte? {
+        return if (isEmpty()) null else maxWith(comparator)
+    }
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableByteArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *